fixing bugs, keeping also groundtrith and soimulation data, new normalization methods (leave one out)

+++ src/methods/ABseqSimulationBenchmark/Benchmark.py
@@ -12,11 +12,17 @@ import subprocess
 import os
 import shutil
 
+def removesuffix(word, suffix):
+    if(word.endswith(suffix)):
+        wordNew = word[:-len(suffix)]
+        word = wordNew
+    return(word)
+
 class Parameters():
 
     iniFile = ""
     normMethods=[]
-    benchmark=False
+    batchEffect=False
     datasets = ""
 
     def __parseParameters(self, paramFile):
@@ -30,10 +36,10 @@ class Parameters():
                 info = str(info[1]).split(",")
                 for element in info:
                     self.normMethods.append(element)
-            elif(str.startswith(line, "benchmark")):
-                info = re.match(("benchmark=(.*)"), line)
+            elif(str.startswith(line, "batchEffect")):
+                info = re.match(("batchEffect=(.*)"), line)
                 if(int(info[1]) == 1):
-                    self.benchmark = True
+                    self.batchEffect = True
 
     def __parseDatasetDir(self):
         settingsFile = "./settings.ini"
@@ -57,7 +63,7 @@ class Benchmark():
         self.parameters = parameters
 
     def __generateNormalizationIni(self, normOriginFilePath):
-        iniFile = normOriginFilePath.strip('.tsv') + ".ini"
+        iniFile = removesuffix(normOriginFilePath,'.tsv') + ".ini"
         
         if os.path.exists(iniFile):
             os.remove(iniFile)
@@ -71,7 +77,7 @@ class Benchmark():
         iniStream.write("ab_type=ab_type\n")
 
         iniStream.write("removeBatchEffect=")
-        if(self.parameters.benchmark == True):
+        if(self.parameters.batchEffect == True):
             iniStream.write(str(1))
         else:
             iniStream.write(str(0))
@@ -86,38 +92,48 @@ class Benchmark():
         subprocess.run(["./source", "scRNAseq/bin/activate"], shell=True)
 
         #call simulations
-        subprocess.run(["python3 ./src/methods/ABseqSimulation/main.py " + self.parameters.iniFile], shell = True)
+        subprocess.run(["python3 ./src/methods/ABseqSimulation/main.py " + self.parameters.iniFile], shell = True, check = True)
         
         #copy simulations into normnalization folder:
         #from bin/SIMMULATIONS to datasets/
         simulationFilePath = "./bin/SIMULATIONS/"
 
-        simulationName = os.path.basename(self.parameters.iniFile.rstrip('.ini'))
+        simulationName = os.path.basename(removesuffix(self.parameters.iniFile, '.ini'))
         simulatedFileName = simulationName + "_SIMULATED.tsv"
         
         normOriginFilePath = self.parameters.datasets + "/" + simulationName + ".tsv"
         simulationResultFilePath = simulationFilePath + simulatedFileName
 
         commandString = "cp " + simulationResultFilePath + " " + normOriginFilePath
-        subprocess.run([commandString], shell = True)
+        subprocess.run([commandString], shell = True, check = True)
         
         #generate a ini file next to it
         self.__generateNormalizationIni(normOriginFilePath)
 
         #run normalizations
+        if( not self.parameters.normMethods ):
+            print("No normalization methods given!")
+            return
         for norm in self.parameters.normMethods:
             if(norm == "GRAPH"):
                 commandString = "python3 src/methods/GraphNormalization/main.py " + normOriginFilePath
-                subprocess.run([commandString], shell = True)
+                subprocess.run([commandString], shell = True, check = True)
             else:
                 commandString = "/usr/local/bin/Rscript --quiet /Users/t.stohn/Desktop/Normalization/PIPELINE/scRNAseq_Normalization/src/normalization/NormalizationScript.R " + norm + " " + simulationName + ".tsv"
-                subprocess.run([commandString], shell = True)
+                subprocess.run([commandString], shell = True, check = True)
         #move also ground truth into normalization folder
-        groundTruthName = os.path.basename(self.parameters.iniFile.rstrip('.ini')) + "_GROUNDTRUTH.tsv"
+        groundTruthName = os.path.basename(removesuffix(self.parameters.iniFile, '.ini')) + "_GROUNDTRUTH.tsv"
         groundTruthResultFilePath = simulationFilePath + groundTruthName
         commandString = "cp " + groundTruthResultFilePath + " ./bin/NORMALIZED_DATASETS/" + simulationName + "/" + groundTruthName
-        subprocess.run([commandString], shell = True)
+        subprocess.run([commandString], shell = True, check = True)
+        #move also simulated file into normalization folder
+        simulatedName = os.path.basename(removesuffix(self.parameters.iniFile, '.ini')) + "_SIMULATED.tsv"
+        simulatedResultFilePath = simulationFilePath + simulatedName
+        commandString = "cp " + simulatedResultFilePath + " ./bin/NORMALIZED_DATASETS/" + simulationName + "/" + simulatedName
+        subprocess.run([commandString], shell = True, check = True)
+
         #remove the ab_count with ab_count_normalized to still run benchmark on it
+        #from ground truth
         groundTruthFile = ("./bin/NORMALIZED_DATASETS/" + simulationName + "/" + groundTruthName)
         groundTruthDataStream = open(groundTruthFile, "rt")
         dataGroundTruth = groundTruthDataStream.read()
@@ -126,15 +142,25 @@ class Benchmark():
         groundTruthDataStream = open(groundTruthFile, "wt")
         groundTruthDataStream.write(dataGroundTruth)
         groundTruthDataStream.close()
+        #from simulated file
+        simulatedFile = ("./bin/NORMALIZED_DATASETS/" + simulationName + "/" + simulatedName)
+        simulatedDataStream = open(simulatedFile, "rt")
+        dataSimulated = simulatedDataStream.read()
+        dataSimulated = dataSimulated.replace('ab_count', 'ab_count_normalized')
+        simulatedDataStream.close()
+        simulatedDataStream = open(simulatedFile, "wt")
+        simulatedDataStream.write(dataSimulated)
+        simulatedDataStream.close()
 
         #run normalization benchmark
         normResultFilePath = "./bin/NORMALIZED_DATASETS/" + simulationName
         benchmarkCommand = "python3 src/benchmark/benchmark.py " + normResultFilePath
-        subprocess.run([benchmarkCommand], shell = True)
+        subprocess.run([benchmarkCommand], shell = True, check = True)
 
     #every run of a simulation -> normalizaitons -> benchmark results in a folder in bin/BENCHMARK
     #all those folers are beeing stored in a foler called SIMULATIONS_dateTime to keep better track
     def moveIntoOneFolder(self, newSimulationDir):
-        simulationName = os.path.basename(self.parameters.iniFile.rstrip('.ini'))
-        shutil.move("./bin/BENCHMARKED_DATASETS/" + simulationName, newSimulationDir)
+        simulationName = os.path.basename(removesuffix(self.parameters.iniFile, '.ini'))
+        if(os.path.exists("./bin/BENCHMARKED_DATASETS/" + simulationName)):
+            shutil.move("./bin/BENCHMARKED_DATASETS/" + simulationName, newSimulationDir)
